requirements.txt
easyocr==1.7.1
matplotlib==3.9.2
numpy==2.1.1
opencv_python==4.10.0.84
pdf2image==1.17.0
Pillow==10.4.0
pytesseract==0.3.13
torch==2.4.1+cu118
